Get zonar schedule data for all routes

Use zonar_get_schedules to retrieve the time each bus (asset) reaches a route anchor (zone) from the Zonar API

Usually zonar_get_schedules_for_time and zonar_get_schedules_for_zone will not be called directly, consider using zonar_get_schedules instead.

Usage

zonar_get_schedules(
  start,
  end,
  by = c("time", "zone"),
  timezone = "America/New_York",
  reformat = FALSE,
  include_geometry = FALSE,
  cachedir = NULL,
  omit_categories = NULL,
  test = FALSE
)

zonar_get_schedules_for_time(
  start,
  end,
  timezone = "America/New_York",
  cachedir = NULL
)

zonar_get_schedules_for_zone(
  zoneid,
  start,
  end,
  timezone = "America/New_York",
  cachedir = NULL
)

Arguments

start

Character vector of length one giving the start time in yyyy-mm-dd hh:mm:ss format.

end

Character vector of length one giving the end time in yyyy-mm-dd hh:mm:ss format.

by

Whether to split requests by time or zone under the hood. By default a reasonable value is chosen based on the amount of data you request (i.e., the size of the interval between start and end.

timezone

Defaults to "America/New_York"

reformat

If TRUE cleanup and re-format the data such that each row represents a "Zone event", i.e., a discrete entry into and/or exit from a zone. Otherwise leave the date in the original format returned by Zonar. Defaults to FALSE.

include_geometry

If TRUE also return the geographic polygon showing the location and extent of each Zone. Defaults to FALSE.

cachedir

Ignored, always uses .zonarCache now.

omit_categories

Character vector containing zone categories to skip/omit.

test

Set to TRUE for testing purposes, otherwise it will be too slow.

zoneid

Integer vector of length one giving the Zonar id of the zone to retrieve.

Details

The Zonar schedule report API can give us that information, but not all at once. Using this R wrapper makes it easy to retrieve bulk schedule reports from the Zonar API in a convenient format that can easily be saved to a database or spreadsheet.

Note that retrieving data for short intervals (one day or less) will generally be faster when setting by = "time", while bulk retrieval of weeks or months of data will be faster when setting by = "zone". By default zonar_get_shedules_all will try to pick the best setting for you depending on the start and end parameters you give it.

Functions

zonar_get_schedules(): Get zonar schedule data for all routes

zonar_get_schedules_for_time(): Get schedule data for a specific time

zonar_get_schedules_for_zone(): Get schedule data for a specific zone

Examples

library(RZonar)
## use by="time" for short duration, e.g. < 1 day
schedtest1 <- zonar_get_schedules(
  start = "2022-11-30 08:00:00",
  end = "2022-11-30 08:30:00",
  by = "time")
dplyr::glimpse(schedtest1)
#> Rows: 2,538
#> Columns: 11
#> $ zoneID             <int> 2272, 2272, 2272, 2272, 2290, 2290, 2290, 2290, 229…
#> $ Zone               <chr> "100 Desoto Rd", "100 Desoto Rd", "100 Desoto Rd", …
#> $ category           <chr> "WATCH", "WATCH", "WATCH", "WATCH", "WATCH", "WATCH…
#> $ `Asset ID`         <chr> "1521", "1521", "161", "161", "1051", "1051", "538"…
#> $ Asset              <chr> "HS521", "HS521", "HS381", "HS381", "B433", "B433",…
#> $ Time               <dttm> 2022-11-30 08:11:17, 2022-11-30 08:12:05, 2022-11-…
#> $ `IN/OUT`           <chr> "IN", "OUT", "IN", "OUT", "IN", "OUT", "IN", "OUT",…
#> $ `Time In Zone`     <chr> NA, "00:00:48", NA, "00:00:30", NA, "00:00:36", NA,…
#> $ Duration           <chr> "00:00:00", "00:00:48", "00:00:00", "00:00:30", "00…
#> $ `Duration Total`   <chr> "00:00:00", "00:00:48", "00:00:00", "00:00:30", "00…
#> $ `Distance (Miles)` <chr> "0.0", "0.1", "0.0", "0.1", "0.0", "0.1", "0.0", "0…

## use by="zone" for bulk retrieval (e.g. weeks or months of data)
schedtest2 <- zonar_get_schedules(
  start = "2022-11-30 08:00:00",
  end = "2022-11-30 08:30:00",
  by = "zone",
  test = TRUE # by zone too slow otherwise
  )
#> NOTE: only pulling schedule data for the first five zones; omit 'test' parameter if you want everything
dplyr::glimpse(schedtest2)
#> Rows: 25
#> Columns: 11
#> $ zoneID             <int> 2272, 2272, 2272, 2272, 2290, 2290, 2290, 2290, 229…
#> $ Zone               <chr> "100 Desoto Rd", "100 Desoto Rd", "100 Desoto Rd", …
#> $ category           <chr> "WATCH", "WATCH", "WATCH", "WATCH", "WATCH", "WATCH…
#> $ Asset              <chr> "HS381", "HS381", "HS521", "HS521", "B433", "B433",…
#> $ `Asset ID`         <chr> "161", "161", "1521", "1521", "1051", "1051", "120"…
#> $ Time               <dttm> 2022-11-30 08:26:46, 2022-11-30 08:27:16, 2022-11-…
#> $ `IN/OUT`           <chr> "IN", "OUT", "IN", "OUT", "IN", "OUT", "IN", "OUT",…
#> $ `Time In Zone`     <chr> NA, "00:00:30", NA, "00:00:48", NA, "00:00:36", NA,…
#> $ Duration           <chr> "00:00:00", "00:00:30", "00:00:00", "00:00:48", "00…
#> $ `Duration Total`   <chr> "00:00:00", "00:00:30", "00:00:00", "00:00:48", "00…
#> $ `Distance (Miles)` <chr> "0.0", "0.1", "0.0", "0.1", "0.0", "0.1", "0.0", "0…

schedtest <- zonar_get_schedules_for_time(
  start = "2022-11-30 08:00:00", # the interval between start and end must be < 15
  end = "2022-11-30 08:14:00"  # minutes, otherwise use zonar_get_schedules
  )
dplyr::glimpse(schedtest)
#> Rows: 1,053
#> Columns: 9
#> $ `Asset ID`         <chr> "1521", "1521", "1051", "1051", "538", "538", "747"…
#> $ Asset              <chr> "HS521", "HS521", "B433", "B433", "HS312", "HS312",…
#> $ Time               <dttm> 2022-11-30 08:11:17, 2022-11-30 08:12:05, 2022-11-…
#> $ `IN/OUT`           <chr> "IN", "OUT", "IN", "OUT", "IN", "OUT", "IN", "OUT",…
#> $ Zone               <chr> "100 Desoto Rd", "100 Desoto Rd", "100 S Huntington…
#> $ `Time In Zone`     <chr> NA, "00:00:48", NA, "00:00:36", NA, "00:00:37", NA,…
#> $ Duration           <chr> "00:00:00", "00:00:48", "00:00:00", "00:00:36", "00…
#> $ `Duration Total`   <chr> "00:00:00", "00:00:48", "00:00:00", "00:00:36", "00…
#> $ `Distance (Miles)` <chr> "0.0", "0.1", "0.0", "0.1", "0.0", "0.1", "0.0", "0…

schedtest <- zonar_get_schedules_for_zone(
  zoneid = 291,
  start = "2022-11-30 08:00:00",
  end = "2022-11-30 09:00:00")
dplyr::glimpse(schedtest)
#> Rows: 12
#> Columns: 9
#> $ Asset              <chr> "HS213", "HS213", "HS225", "HS225", "HS251", "HS251…
#> $ `Asset ID`         <chr> "79", "79", "1282", "1282", "1442", "1442", "271", …
#> $ Time               <dttm> 2022-11-30 08:02:26, 2022-11-30 08:03:25, 2022-11-…
#> $ `IN/OUT`           <chr> "IN", "OUT", "IN", "OUT", "IN", "OUT", "IN", "OUT",…
#> $ `Time In Zone`     <chr> NA, "00:00:59", NA, "00:00:47", NA, "00:00:49", NA,…
#> $ Duration           <chr> "00:00:00", "00:00:59", "00:00:00", "00:00:47", "00…
#> $ `Duration Total`   <chr> "00:00:00", "00:00:59", "00:00:00", "00:00:47", "00…
#> $ `Distance (Miles)` <chr> "0.0", "0.2", "0.0", "0.2", "0.0", "0.2", "0.0", "0…
#> $ zoneID             <dbl> 291, 291, 291, 291, 291, 291, 291, 291, 291, 291, 2…